These interface allways send deprecated because of wrong inheritance
(cherry picked from commit 02fe171b7cba5a91a4540912ea3086e259c61257)
